don't generate thumbnails when listing files on backend side
if it already exists, use it directly.
else, return a different JWT that contains the request path (don't return the
real path to avoid leaking info) and the thumbnail will be generated when
accessing `/thumb/<token>`.

However, this might leave empty spaces and broken images if a thumbnail
can't be generated when accessing its `/thumb/` URL.